[#24] Feat: 페이지 URL 구조 변경
