+++ pages/[category]/[slug]/index.tsx
